Task_2_Edge_VM/requirements.txt
paho-mqtt>=2.0.0
pika
